fix alias paths and move shared files
