+++ app/src/components/ui/dashboard-content.tsx
@@ -1,9 +1,9 @@
-import GoalPage from '../pages/GoalPage';
+import GoalPage from '../../pages/GoalPage';
 import { useAppShell } from './app-shell-context';
-import ProjectPage from '../pages/ProjectPage';
-import HabitPage from '../pages/HabitPage';
-import SettingPage from '../pages/SettingPage';
-import HomePage from '../pages/HomePage';
+import ProjectPage from '../../pages/ProjectPage';
+import HabitPage from '../../pages/HabitPage';
+import SettingPage from '../../pages/SettingPage';
+import HomePage from '../../pages/HomePage';
 
 export default function DashboardContent() {  
   const { activeTab } = useAppShell();
