+++ app/tsconfig.json
@@ -18,6 +18,6 @@
       "@shared/*": ["../shared/*"]
     }
   },
-  "include": ["src/**/*", "**/*.ts", "**/*.tsx", "vite.config.mts"],
+  "include": ["src/**/*", "**/*.ts", "**/*.tsx", "../shared/**/*", "vite.config.mts"],
   "exclude": ["node_modules"]
 }
