+++ server/tsconfig.json
@@ -10,5 +10,5 @@
       "@shared/*": ["../shared/*"]
     }
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*", "../shared/**/*"]
 }
