+++ shared/models/Chat.ts
