+++ shared/models/Topic.ts
