+++ shared/utils/fileConversion.ts
